+++ devel/setup/3-java17.sh
@@ -4,13 +4,15 @@ jdkV=17
 sudo dnf -y install java-$jdkV-openjdk-devel
 
 mavenV=3.9.2
-file=apache-maven-$mavenV-bin.tar.gz
-url=https://dlcdn.apache.org/maven/maven-3/$mavenV/binaries/$file
+dir=apache-maven-$mavenV
+file=$dir-bin.tar.gz
+url=https://dlcdn.apache.org/maven/maven-3/$mavenV/binaries
 
 rm -f $file
 wget $url/$file
-tar -xvf $file
-rm $file
-
-
+sudo mv $file /opt/.
+cd /opt
+sudo tar -xf $file
+sudo rm $file
+ls -l $dir